+++ samples/usage/remote/server.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env ts-node
 
-import { SimplePromptbookLibrary } from '@promptbook/core';
+import { createPromptbookLibraryFromSources } from '@promptbook/core';
 import { OpenAiExecutionTools } from '@promptbook/openai';
 import { runRemoteServer } from '@promptbook/remote-server';
 import chalk from 'chalk';
@@ -19,7 +19,7 @@ main();
 async function main() {
     console.info(chalk.bgGray('🔵 Testing remote server of PromptBook'));
 
-    const library = SimplePromptbookLibrary.fromSources(
+    const library = createPromptbookLibraryFromSources(
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         (await readFile('./samples/templates/50-advanced.ptbk.md', 'utf-8')) as any,
     );