+++ src/library/constructors/createPromptbookLibraryFromList.ts
@@ -1,0 +1,10 @@
+import { PromptbookLibrary } from '../PromptbookLibrary';
+
+export function createPromptbookLibraryFromList(): PromptbookLibrary {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return null as any;
+}
+
+/***
+ * TODO: !!! Annotate all + all to README and samples
+ */